Gantt-like Diagram to plot Event's Time Extension

Gantt-like Diagram to plot Event's Time Extension

Usage

ganttPlotEvents(events, add = FALSE, y1 = 1, y2 = y1 + 1, xlim = NULL,
  ylim = c(min(y1), max(y2)), col = "black", density = 5,
  showLabels = TRUE, eventLabels = rownames(events), yLabel = (y1 + y2)/2,
  type = "rectangle", title = "", leftMargin = 0.2, xlab = "Time",
  cex = 0.8, indicate = NULL, indicationColour = "red",
  bandheight = 0.1, alternating = FALSE, adj = 0.5, ...)

Arguments

events

event list as retrieved by hsEvents. Required columns: tBeg (begin of event) and tEnd (end of event), both of class POSIXt

add

if TRUE, the event boxes are added to the current plot, otherwise a new plot is generated

y1

lower coordinates of the event boxes

y2

upper coordinates of the event boxes

xlim

x limits. If NULL (default) the limits will be chosen so that all events fit into the plot

ylim

y limits

col

colour of shading lines

density

density of shading lines

showLabels

if TRUE, the event boxes are labelled with the row names of the events

eventLabels

labels to be given to the events. Default: rownames(events)

yLabel

y-position of labels, if labels are to be shown

type

one of c("rectange", "vertical")

title

title to be plotted left of event rectangles

leftMargin

left margin (where title is printed) as fraction of the range of the total time interval spanned by the events

xlab

x axis label

cex

character expansion factor

indicate

indices of events to be indicated in a different color (indicationColuor)

indicationColour

colour to be used for indication, default: "red" extension factor for labels (event numbers)

bandheight

passed to addLabels

alternating

passed to addLabels

adj

passed to text plotting the event labels

...

further arguments passed to rect or segments